#RoipBackend: 1)Token authorization logic implementation in project(view Program.cs, Utillities package, UserService.LogInAsync(), UserService.LogoutAsync(), some of the Controllers API functions) 2)Adjusted more project's clean Architecture principles 3)Users / Products Page adjustment in Get...Async() Services functions
#RoipFrontend: None

+++ RoipBackend/Controllers/ProductsController.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using RoipBackend.Interfaces;
 using RoipBackend.Models;
 using RoipBackend.Services;
 using RoipBackend.Utilities;
@@ -10,66 +11,77 @@ namespace RoipBackend.Controllers
     [Route("api/[controller]")]
     public class ProductsController : ControllerBase
     {
-        //TODO: Add refresh token / cookies functionality (???)
-
+        //TODO: Add Loggers to all functions
         private readonly IConfiguration _configuration;
         private readonly ProductService _productService;
         private readonly LoggerService _loggerService;
+        private readonly JwtHelper _jwtHelper;
 
-        public ProductsController(IConfiguration configuration, ProductService productService, LoggerService loggerService)
+        public ProductsController(IConfiguration configuration, ProductService productService, LoggerService loggerService, JwtHelper jwtHelper)
         {
             _configuration = configuration;
             _productService = productService;
             _loggerService = loggerService;
+            _jwtHelper = jwtHelper;
         }
 
         [HttpGet("get-all-products")]
         [Authorize(Roles = "Admin,Customer")]
-        public async Task<IActionResult> GetAllProductsAsync()
-        {
-            var serviceResult = await _productService.GetAllProductsAsync();
-            return HandleServiceResult(serviceResult);
+        public async Task<IActionResult> GetAllProductsAsync(string jwt, int pageNumber, int pageSize)
+        {          
+            IActionResult jwt_ValidationResult = _jwtHelper.JwtCheck(jwt);
+            if (jwt_ValidationResult is OkResult)
+                return await _productService.GetAllProductsAsync(pageNumber, pageSize);
+            return jwt_ValidationResult;
         }
 
         [HttpPost("add-product")]
         [Authorize(Roles = "Admin,Customer")]
-        public async Task<IActionResult> AddProductAsync([FromBody] Product product)
+        public async Task<IActionResult> AddProductAsync(string jwt, Product product)
         {
+            IActionResult jwt_ValidationResult = _jwtHelper.JwtCheck(jwt);
+            if (jwt_ValidationResult is UnauthorizedObjectResult)
+                return jwt_ValidationResult;
+
             if (!ModelState.IsValid)                                
                 return ModelStateError(Consts.VALIDATION_FAILED_STR);
             
-            IActionResult serviceResult = await _productService.AddProductAsync(product);
-            return HandleServiceResult(serviceResult);
+            return await _productService.AddProductAsync(product);
         }
 
         [HttpPut("buy-product/{productName}")]
         [Authorize(Roles = "Admin,Customer")]
-        public async Task<IActionResult> BuyProductAsync(string productName, [FromBody] int quantity)
+        public async Task<IActionResult> BuyProductAsync(string jwt, string productName, int quantity)
         {
+            IActionResult jwt_ValidationResult = _jwtHelper.JwtCheck(jwt);
+            if (jwt_ValidationResult is UnauthorizedObjectResult)
+                return jwt_ValidationResult;
+
             if (!ModelState.IsValid)
                 return ModelStateError(Consts.VALIDATION_FAILED_STR);
 
-            IActionResult serviceResult = await _productService.BuyProductAsync(productName, quantity);
-            return HandleServiceResult(serviceResult);
+            return await _productService.BuyProductAsync(productName, quantity);
         }
 
         [HttpDelete("delete-product/{productName}")]
         [Authorize(Roles = "Admin")]
-        public async Task<IActionResult> DeleteProductAsync(string productName)
+        public async Task<IActionResult> DeleteProductAsync(string jwt, string productName)
         {
+            IActionResult jwt_ValidationResult = _jwtHelper.JwtCheck(jwt);
+            if (jwt_ValidationResult is UnauthorizedObjectResult)
+                return jwt_ValidationResult;
+
             if (!ModelState.IsValid)
                 return ModelStateError(Consts.VALIDATION_FAILED_STR);
 
-            IActionResult serviceResult = await _productService.DeleteProductAsync(productName);
-            return HandleServiceResult(serviceResult);
+            return await _productService.DeleteProductAsync(productName);
         }
 
         [HttpGet("search-filter")]
         [Authorize(Roles = "Admin,Customer")]
-        public async Task<IActionResult> SearchFilterAsync([FromQuery] string filterText)
+        public async Task<IActionResult> SearchFilterAsync(string filterText)
         {
-            IActionResult serviceResult = await _productService.SearchFilterAsync(filterText);
-            return HandleServiceResult(serviceResult);
+            return await _productService.SearchFilterAsync(filterText);
         }
 
         private BadRequestObjectResult ModelStateError(string message)
@@ -79,12 +91,10 @@ namespace RoipBackend.Controllers
                 .Select(e => e.ErrorMessage)
                 .ToList();
 
-            return new BadRequestObjectResult(new { Message = message, Errors = errors });
-        }
-
-        private static ActionResult HandleServiceResult(IActionResult serviceResult)
-        {
-            return ServiceResultHandler.HandleServiceResult(serviceResult);
-        }        
+            return new BadRequestObjectResult(new { Message = message, Errors = errors })
+            {
+                StatusCode = StatusCodes.Status400BadRequest
+            };
+        }       
     }
 }
